refactor: remove redundant casts (Rector, strictly safe subset)
Behaviour-preserving cleanup only, hand-picked rules (15 files):
- RecastingRemovalRector — drop casts already matching the value's type
- RemoveConcatAutocastRector — drop (string) casts redundant in concatenation
- RemoveUnusedNonEmptyArrayBeforeForeachRector — drop redundant if-guard before foreach

RemoveUnusedVariableAssignRector was evaluated and REJECTED: it dropped the
assignment of a security filter's return value in PageController
(`$relatedAlbums = $this->filterAlbumsByAccess(...)` -> bare call, which
discards the filtered result and would leak private/NSFW related albums), so
it is excluded. The broader Rector sets (readonly, switch->match, explicit-bool
compare, first-class-callable, constructor promotion) remain deferred to a
separately-reviewed pass.

Validated: PHPUnit 84 green; PHPStan delta zero; style re-checked with the
formatter.

+++ app/Controllers/Admin/CollectionsController.php
@@ -311,7 +311,7 @@ final class CollectionsController extends BaseController
             'total' => $total,
             'page' => $page,
             'per_page' => self::PICKER_PER_PAGE,
-            'total_pages' => (int) max(1, (int) ceil($total / self::PICKER_PER_PAGE)),
+            'total_pages' => max(1, (int) ceil($total / self::PICKER_PER_PAGE)),
         ]);
     }
